rdf-to-pg.ts
This change pushed in the src folder many of the utility functions that were developped as extension of the DatasetCore interface
