+++ README.md
@@ -43,7 +43,7 @@ works by connecting to the Gremlin API and making requests to extract the graph.
 
 **RDF -> PG**
 
-- `rdf-to-pg.js` allows to come back to Property Graphs from a RDF previously
+- `rdf-to-pg.ts` allows to come back to Property Graphs from a RDF previously
 generated by `prec` or `neo4j-to-rdf`.
 
 **Tools**
